requirements.txt
amqp==2.4.2
astroid==2.2.5
billiard==3.6.0.0
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Caching==1.7.1
httplib2==0.12.3
idna==2.8
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.5.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.0.1
pkg-resources==0.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pylint==2.3.1
python-dotenv==0.10.2
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
typed-ast==1.3.5
uritemplate==3.0.0
urllib3==1.24.3
vine==1.3.0
visitor==0.1.3
Werkzeug==0.15.2
wrapt==1.11.1
